+++ crates/uselesskey-rustls-der/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "uselesskey-rustls-der"
+version = "0.3.0"
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+description = "rustls-pki-types DER conversion adapters for uselesskey fixtures."
+categories.workspace = true
+keywords = ["rustls", "der", "fixtures", "testing", "adapter"]
+readme = "README.md"
+exclude = ["fuzz/**", "corpus/**", "**/*.der", "**/*.pem"]
+homepage.workspace = true
+documentation = "https://docs.rs/uselesskey-rustls-der"
+authors.workspace = true
+
+[dependencies]
+rustls-pki-types = "1"
+
+uselesskey-x509 = { path = "../uselesskey-x509", version = "0.3.0", optional = true }
+uselesskey-rsa = { path = "../uselesskey-rsa", version = "0.3.0", optional = true }
+uselesskey-ecdsa = { path = "../uselesskey-ecdsa", version = "0.3.0", optional = true }
+uselesskey-ed25519 = { path = "../uselesskey-ed25519", version = "0.3.0", optional = true }
+
+[features]
+default = ["x509"]
+x509 = ["dep:uselesskey-x509"]
+rsa = ["dep:uselesskey-rsa"]
+ecdsa = ["dep:uselesskey-ecdsa"]
+ed25519 = ["dep:uselesskey-ed25519"]
+all = ["x509", "rsa", "ecdsa", "ed25519"]
+
+[dev-dependencies]
+uselesskey-core = { path = "../uselesskey-core", version = "0.3.0" }
+uselesskey-x509 = { path = "../uselesskey-x509", version = "0.3.0" }
+uselesskey-rsa = { path = "../uselesskey-rsa", version = "0.3.0" }
+uselesskey-ecdsa = { path = "../uselesskey-ecdsa", version = "0.3.0" }
+uselesskey-ed25519 = { path = "../uselesskey-ed25519", version = "0.3.0" }
