+++ crates/uselesskey-rustls-der/README.md
@@ -1,0 +1,6 @@
+# uselesskey-rustls-der
+
+DER conversion extension traits for using uselesskey fixtures with `rustls-pki-types`.
+
+This crate focuses only on shape conversions (`PrivateKeyDer` / `CertificateDer`) and is
+used by `uselesskey-rustls`.
